correções do financeiro (negativo e positivo)

+++ database/migrations/2021_07_12_190805_create_products_proposals_table.php
@@ -4,27 +4,26 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateProductsProposalsTable extends Migration
-{
+class CreateProductsProposalsTable extends Migration {
+
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up()
-    {
+    public function up() {
         Schema::create('products_proposals', function (Blueprint $table) {
-			$table->id();
-			$table->foreignId('proposal_id');
-			$table->foreignId('product_id');
-			$table->integer('amount');
-			$table->decimal('subtotalHours', 5, 1)->nullable();
-			$table->integer('subtotalDeadline')->nullable();
-			$table->integer('subtotalCost')->nullable();
-			$table->integer('subtotalTax_rate')->nullable();
-			$table->integer('subtotalPrice');
-			$table->integer('subtotalMargin')->nullable();
-			$table->timestamps();
+            $table->id();
+            $table->foreignId('proposal_id');
+            $table->foreignId('product_id');
+            $table->integer('amount');
+            $table->decimal('subtotalHours', 5, 1)->nullable();
+            $table->integer('subtotalDeadline')->nullable();
+            $table->integer('subtotalCost')->nullable();
+            $table->integer('subtotalTax_rate')->nullable();
+            $table->decimal('subtotalPrice', 8, 2);
+            $table->decimal('subtotalMargin', 8, 2)->nullable();
+            $table->timestamps();
         });
     }
 
@@ -33,8 +32,8 @@ class CreateProductsProposalsTable extends Migration
      *
      * @return void
      */
-    public function down()
-    {
+    public function down() {
         Schema::dropIfExists('products_proposals');
     }
+
 }
